+++ docs/source/conf.py
@@ -24,7 +24,7 @@ copyright = "2020, Alex Malins"
 author = "Alex Malins"
 
 # The full version, including alpha/beta/rc tags
-release = "0.0.7"
+release = "0.0.8"
 
 
 # -- General configuration ---------------------------------------------------