+++ radioactivedecay/decayfunctions.py
@@ -1,7 +1,6 @@
 """
 The decacyfunctions module defines the two main classes used by users: the ``Inventory`` and
-``Radionuclide`` classes. It also contains functions to parse radionuclide strings, check the
-validity of inventory dictionaries, and convert between time units.
+``Radionuclide`` classes.
 
 The examples shown assume the ``radioactivedecay`` package has been imported as:
 
@@ -16,232 +15,17 @@ DEFAULTDATA : DecayData
     Default decay dataset used by ``radioactivedecay``. This is currently ICRP 107.
 """
 
-from functools import singledispatch, update_wrapper
 import numpy as np
 from radioactivedecay.decaydata import DecayData
+from radioactivedecay.utils import (
+    parse_radionuclide,
+    check_dictionary,
+    time_unit_conv,
+    add_dictionaries,
+    method_dispatch,
+)
 
 DEFAULTDATA = DecayData("icrp107")
-
-
-def parse_nuclide_name(nuclide_name, nuclide_names, dataset):
-    """
-    Parses a radionuclide string and checks whether the radionuclide is contained in the decay
-    dataset.
-
-    Parameters
-    ----------
-    nuclide_name : str
-        Radionuclide string.
-    nuclide_names : numpy.ndarray
-        NumPy array of all the radionuclides in the decay dataset.
-    dataset : str
-        Name of the decay dataset.
-
-    Returns
-    -------
-    str
-        Radionuclide string parsed in symbol - mass number format.
-
-    Raises
-    ------
-    ValueError
-        If the radionuclide string is invalid or it is not contained in the decay dataset.
-
-    Examples
-    --------
-    >>> rd.parse_nuclide_name('222Rn', rd.DEFAULTDATA.nuclide_names, rd.DEFAULTDATA.dataset)
-    'Rn-222'
-    >>> rd.parse_nuclide_name('Ba137m', rd.DEFAULTDATA.nuclide_names, rd.DEFAULTDATA.dataset)
-    'Ba-137m'
-
-    """
-
-    letter_flag, number_flag = False, False
-    for char in nuclide_name:
-        if char.isalpha():
-            letter_flag = True
-        if char.isdigit():
-            number_flag = True
-        if letter_flag and number_flag:
-            break
-
-    if (
-        not (letter_flag and number_flag)
-        or len(nuclide_name) < 2
-        or len(nuclide_name) > 7
-    ):
-        raise ValueError(
-            str(nuclide_name)
-            + " is not a valid radionuclide in"
-            + dataset
-            + " dataset."
-        )
-
-    original_nuclide_name = nuclide_name
-
-    while nuclide_name[0].isdigit():  # Re-order inputs e.g. 99mTc to Tc99m.
-        nuclide_name = nuclide_name[1:] + nuclide_name[0]
-    if nuclide_name[0] in ["m", "n"]:
-        nuclide_name = nuclide_name[1:] + nuclide_name[0]
-
-    for i in range(1, len(nuclide_name)):  # Add hyphen e.g. Tc99m to Tc-99m.
-        if nuclide_name[i].isdigit():
-            if nuclide_name[i - 1] != "-":
-                nuclide_name = nuclide_name[:i] + "-" + nuclide_name[i:]
-            break
-
-    if nuclide_name not in nuclide_names:
-        raise ValueError(
-            str(original_nuclide_name)
-            + " is not a valid radionuclide in "
-            + dataset
-            + " dataset."
-        )
-
-    return nuclide_name
-
-
-def check_dictionary(inv_dict, nuclide_names, dataset):
-    """
-    Checks validity of a dictionary of radionuclides and associated acitivities. Radionuclides
-    must be in the decay dataset.
-
-    Parameters
-    ----------
-    inv_dict : dict
-        Dictionary containing radionuclide strings as keys and activities as values.
-    nuclide_names : list or numpy.ndarray
-        NumPy array of all the radionuclides in the decay dataset.
-    dataset : str
-        Name of the decay dataset.
-
-    Returns
-    -------
-    dict
-        Dictionary where the contents have been validated and the radionuclide key strings have
-        been parsed into symbol - mass number format.
-
-    Raises
-    ------
-    ValueError
-        If an activity key is invalid.
-
-    Examples
-    --------
-    >>> rd.check_dictionary({'3H': 1.0}, rd.DEFAULTDATA.nuclide_names, rd.DEFAULTDATA.dataset)
-    {'H-3': 1.0}
-
-
-    """
-
-    inv_dict = {
-        parse_nuclide_name(nuc, nuclide_names, dataset): act
-        for nuc, act in inv_dict.items()
-    }
-    for nuc, act in inv_dict.items():
-        if not isinstance(act, (float, int)):
-            raise ValueError(
-                str(act) + " is not a valid radioactivity for " + str(nuc) + "."
-            )
-
-    return inv_dict
-
-
-def time_unit_conv(time, units_from, units_to, year_conv):
-    """
-    Converts a time from one set of units to another.
-
-    Parameters
-    ----------
-    time : float
-        Time before conversion.
-    units_from : str
-        Time unit before conversion
-    units_to : str
-        Time unit after conversion
-    yeav_conv : float or int
-        Conversion factor for number of days in a year.
-
-    Returns
-    -------
-    float
-        Time in new units.
-
-    Raises
-    ------
-    ValueError
-        If one of the time unit parameters is invalid.
-
-    Examples
-    --------
-    >>> rd.time_unit_conv(1.0, 'd', 'h', rd.DEFAULTDATA.year_conv)
-    24.0
-
-    """
-
-    conv = {
-        "ns": 1.0e-9,
-        "us": 1.0e-6,
-        "ms": 1.0e-3,
-        "s": 1.0,
-        "m": 60.0,
-        "h": 3600.0,
-        "d": 86400.0,
-        "y": 86400.0 * year_conv,
-        "sec": 1,
-        "second": 1,
-        "seconds": 1,
-        "hr": 3600.0,
-        "hour": 3600.0,
-        "hours": 3600.0,
-        "day": 86400.0,
-        "days": 86400.0,
-        "yr": 86400.0 * year_conv,
-        "year": 86400.0 * year_conv,
-        "years": 86400.0 * year_conv,
-        "ky": 86400.0 * year_conv * 1.0e3,
-        "My": 86400.0 * year_conv * 1.0e6,
-        "Gy": 86400.0 * year_conv * 1.0e9,
-        "Ty": 86400.0 * year_conv * 1.0e12,
-        "Py": 86400.0 * year_conv * 1.0e15,
-    }
-
-    if units_from not in conv:
-        raise ValueError(
-            str(units_from) + ' is not a valid unit, e.g. "s", "m", "h", "d" or "y".'
-        )
-    if units_to not in conv:
-        raise ValueError(
-            str(units_to) + ' is not a valid unit, e.g. "s", "m", "h", "d" or "y".'
-        )
-
-    return time * conv[units_from] / conv[units_to]
-
-
-def add_dictionaries(dict1, dict2):
-    """Adds together two dictionaries of radionuclies and associated acitivities."""
-
-    new_dict = dict1.copy()
-    for nuclide, radioactivity in dict2.items():
-        if nuclide in new_dict:
-            new_dict[nuclide] = new_dict[nuclide] + radioactivity
-        else:
-            new_dict[nuclide] = radioactivity
-
-    return new_dict
-
-
-def method_dispatch(func):
-    """Adds singledispatch support for class methods."""
-
-    dispatcher = singledispatch(func)
-
-    def wrapper(*args, **kw):
-        return dispatcher.dispatch(args[1].__class__)(*args, **kw)
-
-    wrapper.register = dispatcher.register
-    update_wrapper(wrapper, func)
-    return wrapper
 
 
 class Inventory:
@@ -282,7 +66,7 @@ class Inventory:
         """
 
         if check is True:
-            contents = check_dictionary(contents, data.nuclide_names, data.dataset)
+            contents = check_dictionary(contents, data.radionuclides, data.dataset)
         self.contents = dict(sorted(contents.items(), key=lambda x: x[0]))
         self.data = data
 
@@ -334,7 +118,7 @@ class Inventory:
         """
 
         add_contents = check_dictionary(
-            add_contents, self.data.nuclide_names, self.data.dataset
+            add_contents, self.data.radionuclides, self.data.dataset
         )
         new_contents = add_dictionaries(self.contents, add_contents)
         self.change(new_contents, False, self.data)
@@ -358,7 +142,7 @@ class Inventory:
 
         """
         sub_contents = check_dictionary(
-            sub_contents, self.data.nuclide_names, self.data.dataset
+            sub_contents, self.data.radionuclides, self.data.dataset
         )
         sub_contents.update(
             (nuclide, radioactivity * -1.0)
@@ -455,7 +239,7 @@ class Inventory:
     @remove.register(str)
     def _(self, delete):
         """Remove radionuclide string from this inventory."""
-        delete = parse_nuclide_name(delete, self.data.nuclide_names, self.data.dataset)
+        delete = parse_radionuclide(delete, self.data.radionuclides, self.data.dataset)
         new_contents = self.contents.copy()
         if delete not in new_contents:
             raise ValueError(delete + " does not exist in this inventory.")
@@ -466,7 +250,7 @@ class Inventory:
     def _(self, delete):
         """Remove list of radionuclide(s) from this inventory."""
         delete = [
-            parse_nuclide_name(nuc, self.data.nuclide_names, self.data.dataset)
+            parse_radionuclide(nuc, self.data.radionuclides, self.data.dataset)
             for nuc in delete
         ]
         new_contents = self.contents.copy()
@@ -515,11 +299,11 @@ class Inventory:
             )
         )
 
-        vector_n0 = np.zeros([self.data.no_nuclides], dtype=np.float64)
+        vector_n0 = np.zeros([self.data.no_radionuclides], dtype=np.float64)
         indices = set()
-        for nuclide_name in self.contents:
-            i = self.data.nuclide_dict[nuclide_name]
-            vector_n0[i] = self.contents[nuclide_name] / self.data.decay_consts[i]
+        for radionuclide in self.contents:
+            i = self.data.radionuclide_dict[radionuclide]
+            vector_n0[i] = self.contents[radionuclide] / self.data.decay_consts[i]
             indices.update(self.data.matrix_c[:, i].nonzero()[0])
         indices = list(indices)
 
@@ -532,7 +316,7 @@ class Inventory:
         ).dot(vector_n0)
         vector_at = np.multiply(vector_nt, self.data.decay_consts)
 
-        new_contents = dict(zip(self.data.nuclide_names[indices], vector_at[indices]))
+        new_contents = dict(zip(self.data.radionuclides[indices], vector_at[indices]))
         new_contents = dict(sorted(new_contents.items(), key=lambda x: x[0]))
         return Inventory(new_contents, False, self.data)
 
@@ -548,17 +332,19 @@ class Radionuclide:
 
     Parameters
     ----------
-    nuclide_name : str
+    radionuclide : str
         Radionuclide string.
     data : DecayData, optional
         Decay dataset (default is the ICRP 107 dataset).
 
     Attributes
     ----------
-    nuclide_name : str
+    radionuclide : str
         Radionuclide string.
     decay_constant : numpy.float64
         Decay constant of the radionuclide (s\\ :sup:`-1`).
+    prog_bf : dict
+        Dictionary containing first progeny as keys and branching fractions as values.
     data : DecayData
         Decay dataset.
 
@@ -569,26 +355,29 @@ class Radionuclide:
 
     """
 
-    ln2 = np.log(2)
+    def __init__(self, radionuclide, data=DEFAULTDATA):
+        self.change(radionuclide, data)
 
-    def __init__(self, nuclide_name, data=DEFAULTDATA):
-        self.change(nuclide_name, data)
-
-    def change(self, nuclide_name, data):
+    def change(self, radionuclide, data):
         """
-        Changes or initializes the nuclide_name, decay_constant and data attritubes of this
+        Changes or initializes the radionuclide, decay_constant and data attritubes of this
         Radionuclide instance.
         """
 
-        self.nuclide_name = parse_nuclide_name(
-            nuclide_name, data.nuclide_names, data.dataset
+        self.radionuclide = parse_radionuclide(
+            radionuclide, data.radionuclides, data.dataset
         )
-        self.decay_constant = data.decay_consts[data.nuclide_dict[self.nuclide_name]]
+        self.decay_constant = data.decay_consts[
+            data.radionuclide_dict[self.radionuclide]
+        ]
+        self.prog_bf_mode = data.prog_bfs_modes[
+            data.radionuclide_dict[self.radionuclide]
+        ]
         self.data = data
 
     def half_life(self, units="s"):
         """
-        Returns half-life of radionuclide in chosen units.
+        Returns half-life of the radionuclide in chosen units.
 
         Parameters
         ----------
@@ -617,12 +406,70 @@ class Radionuclide:
                 1.0, units_from="s", units_to=units, year_conv=self.data.year_conv
             )
         )
-        return conv * Radionuclide.ln2 / self.decay_constant
+        return conv * self.data.ln2 / self.decay_constant
+
+    def progeny(self):
+        """
+        Returns the first progeny of the radionuclide.
+
+        Returns
+        -------
+        list
+            List of the first progeny of the radionuclide, ordered by decreasing branching
+            fraction.
+
+        Examples
+        --------
+        >>> K40 = rd.Radionuclide('K-40')
+        >>> K40.progeny()
+        ['Ca-40', 'Ar-40']
+
+        """
+
+        return list(self.prog_bf_mode.keys())
+
+    def branching_fractions(self):
+        """
+        Returns the branching fractions for the first progeny of the radionuclide.
+
+        Returns
+        -------
+        list
+            List of branching fractions.
+
+        Examples
+        --------
+        >>> K40 = rd.Radionuclide('K-40')
+        >>> K40.branching_fractions()
+        [0.8914, 0.1086]
+
+        """
+
+        return [bf_mode[0] for bf_mode in list(self.prog_bf_mode.values())]
+
+    def decay_modes(self):
+        """
+        Returns the decay modes creating the first progeny of the radionuclide.
+
+        Returns
+        -------
+        list
+            List of decay modes.
+
+        Examples
+        --------
+        >>> K40 = rd.Radionuclide('K-40')
+        >>> K40.decay_modes()
+        ['\u03b2-', '\u03b2+ & EC']
+
+        """
+
+        return [bf_mode[1] for bf_mode in list(self.prog_bf_mode.values())]
 
     def __repr__(self):
         return (
             "Radionuclide: "
-            + str(self.nuclide_name)
+            + str(self.radionuclide)
             + ", Decay dataset: "
             + self.data.dataset
         )